requirements.txt
streamlit==1.22.0
PyPDF2==3.0.1
fpdf2==2.7.4